Логика создания кнопок добавить и удалить переброшена в другой файл
